# api/openapi.yaml
openapi: "3.1.0"
info:
  title: mtix API
  description: |
    REST API for mtix — an AI-native micro issue manager for code-generating LLMs.
    Designed for safety-critical environments (DO-178C, IEC 62304, NASA-STD-8739.8).

    Agent documentation: `.mtix/docs/AGENTS.md`
  version: "1.0.0"
  contact:
    name: mtix
  license:
    name: MIT
servers:
  - url: "http://localhost:{port}"
    description: Local mtix server
    variables:
      port:
        default: "6849"
        description: Configurable server port
paths:
...
  /api/v1/nodes:
    get:
      operationId: listNodes
      summary: List nodes with filters
      tags: [Nodes]
      parameters:
        - $ref: "#/components/parameters/status"
        - $ref: "#/components/parameters/under"
        - $ref: "#/components/parameters/nodeType"
        - $ref: "#/components/parameters/priority"
        - $ref: "#/components/parameters/assignee"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated node list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeList"

    post:
      operationId: createNode
      summary: Create a node
      tags: [Nodes]
      parameters:
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeRequest"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/nodes/{id}:
    get:
      operationId: getNode
      summary: Get node by ID
      tags: [Nodes]
      parameters:
        - $ref: "#/components/parameters/nodeId"
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateNode
      summary: Update node fields
      tags: [Nodes]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteNode
      summary: Soft-delete node and descendants
      tags: [Nodes]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
        - name: cascade
          in: query
          schema:
            type: boolean
            default: true
          description: If false, delete only target node (children become orphans)
      responses:
        "200":
          description: Node soft-deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/nodes/{id}/claim:
    post:
      operationId: claimNode
      summary: Claim node for work
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Node claimed
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/nodes/{id}/unclaim:
    post:
      operationId: unclaimNode
      summary: Release claim
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Claim released
...
  /api/v1/nodes/{id}/defer:
    post:
      operationId: deferNode
      summary: Defer node
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                until:
                  type: string
                  format: date-time
                  description: ISO-8601 date. Omit for indefinite deferral.
      responses:
        "200":
          description: Node deferred
  /api/v1/nodes/{id}/cancel:
    post:
      operationId: cancelNode
      summary: Cancel node
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                cascade:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Node cancelled
  /api/v1/nodes/{id}/reopen:
    post:
      operationId: reopenNode
      summary: Reopen done or cancelled node
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Node reopened
...
  /api/v1/nodes/{id}/rerun:
    post:
      operationId: rerunNode
      summary: Trigger rerun — invalidate descendants
      tags: [Workflow]
      parameters:
        - $ref: "#/components/parameters/nodeId"
        - $ref: "#/components/parameters/csrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy]
              properties:
                strategy:
                  type: string
                  enum: [all, open_only, delete, review]
      responses:
        "200":
          description: Rerun triggered
...
  /api/v1/search:
    get:
      operationId: searchNodes
      summary: Full-text search
      tags: [Query]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/status"
        - $ref: "#/components/parameters/nodeType"
        - $ref: "#/components/parameters/under"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeList"

  # ── Dependencies ────────────────────────────────────────────
...
components:
  parameters:
    nodeId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Dot-notation node ID (e.g., PROJ-42.1.3)
    csrfHeader:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        enum: [mtix]
      description: CSRF protection header required on all mutations per NFR-5.5
    status:
      name: status
      in: query
      schema:
        type: string
        enum: [open, in_progress, blocked, done, deferred, cancelled, invalidated]
    under:
      name: under
      in: query
      schema:
        type: string
      description: Filter to subtree under this node ID
    nodeType:
      name: type
      in: query
      schema:
        type: string
        enum: [story, epic, issue, micro, auto]
    priority:
      name: priority
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
    assignee:
      name: assignee
      in: query
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 500
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
    Node:
      type: object
      required: [id, title, status, created_at, updated_at]
      properties:
        id:
          type: string
          description: Dot-notation ID (e.g., PROJ-42.1.3)
        parent_id:
          type: string
        project:
          type: string
        depth:
          type: integer
        seq:
          type: integer
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 51200
        prompt:
          type: string
          maxLength: 102400
        acceptance:
          type: string
        labels:
          type: array
          items:
            type: string
        node_type:
          type: string
          enum: [story, epic, issue, micro, auto]
        issue_type:
          type: string
          enum: [bug, feature, task, chore, refactor, test, doc]
        priority:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [open, in_progress, blocked, done, deferred, cancelled, invalidated]
        previous_status:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 1
        assignee:
          type: string
        creator:
          type: string
        agent_state:
          type: string
          enum: [idle, working, stuck, done]
        weight:
          type: number
        content_hash:
          type: string
        child_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
        defer_until:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
    NodeList:
      type: object
      required: [nodes, total, has_more]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        total:
          type: integer
        has_more:
          type: boolean
    CreateNodeRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
        parent_id:
          type: string
        project:
          type: string
        description:
          type: string
        prompt:
          type: string
        acceptance:
          type: string
        labels:
          type: array
          items:
            type: string
        node_type:
          type: string
          enum: [story, epic, issue, micro, auto]
        issue_type:
          type: string
          enum: [bug, feature, task, chore, refactor, test, doc]
        priority:
          type: integer
          minimum: 1
          maximum: 5
    UpdateNodeRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        prompt:
          type: string
        acceptance:
          type: string
        labels:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 1
          maximum: 5
        assignee:
          type: string
        code_refs:
          type: array
          items:
            type: object
        commit_refs:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code (NOT_FOUND, INVALID_INPUT, etc.)
              enum:
                - NOT_FOUND
                - ALREADY_EXISTS
                - INVALID_INPUT
                - INVALID_TRANSITION
                - CYCLE_DETECTED
                - CONFLICT
                - ALREADY_CLAIMED
                - NODE_BLOCKED
                - STILL_DEFERRED
                - AGENT_STILL_ACTIVE
                - NO_ACTIVE_SESSION
                - INVALID_CONFIG_KEY
            message:
              type: string
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict (duplicate, already claimed, cycle)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer token for cloud sync mode (Phase 5+). Not required for localhost."